update persistent segtree

+++ code/Math/exGCD.cpp
@@ -1,21 +1,21 @@
 // ax+by=gcd(a,b)
-int ex_gcd(int a, int b, int& x, int& y) {
+ll ex_gcd(ll a, ll b, ll& x, ll& y) {
   if (b == 0) {
     x = 1;
     y = 0;
     return a;
   }
-  int d = ex_gcd(b, a % b, x, y);
-  int temp = x;
+  ll d = ex_gcd(b, a % b, x, y);
+  ll temp = x;
   x = y;
   y = temp - a / b * y;
   return d;
 }
 //the smallest integer solution to ax+by=c;
-bool liEu(int a, int b, int c, int& x, int& y) {
-  int d = ex_gcd(a, b, x, y);
+bool liEu(ll a, ll b, ll c, ll& x, ll& y) {
+  ll d = ex_gcd(a, b, x, y);
   if (c % d != 0) return 0;
-  int k = c / d;
+  ll k = c / d;
   x *= k;
   y *= k;
   return 1;
